refactor: VPC 관련 클래스 패키지 경로 정리 및 Command 패턴 적용
