+++ src/main/java/com/agenticcp/core/domain/cloud/port/outbound/vpc/VpcManagementPort.java
@@ -1,4 +1,4 @@
-package com.agenticcp.core.domain.cloud.port.outbound.aws;
+package com.agenticcp.core.domain.cloud.port.outbound.vpc;
 
 import com.agenticcp.core.domain.cloud.entity.CloudResource;
 import com.agenticcp.core.domain.cloud.port.model.VpcCreateRequest;
@@ -13,26 +13,28 @@ public interface VpcManagementPort {
     
     /**
      * VPC 생성
+     * @param command 생성 명령
+     * @return 생성된 VPC
      */
-    CloudResource createVpc(VpcCreateRequest request);
+    CloudResource createVpc(CreateVpcCommand command);
     
     /**
      * VPC 조회 (단일)
      */
-    Optional<CloudResource> getVpc(ResourceIdentity vpcId);
+    Optional<CloudResource> getVpc(GetVpcCommand command);
     
     /**
      * VPC 목록 조회
      */
-    List<CloudResource> listVpcs(VpcQuery query);
+    List<CloudResource> listVpcs(ListVpcsQuery command);
     
     /**
      * VPC 수정
      */
-    CloudResource updateVpc(ResourceIdentity vpcId, VpcUpdateRequest request);
+    CloudResource updateVpc(UpdateVpcCommand command);
     
     /**
      * VPC 삭제
      */
-    void deleteVpc(ResourceIdentity vpcId);
+    void deleteVpc(DeleteVpcCommand command);
 }
